Converte temperatura de Fahrenheit em Grau Celsius

+++ ConvertFahrenheitParaCelsius.py
@@ -1,3 +1,3 @@
-tempFahrenheit = int(input('Informe a temperatura em Fahrenheit: '))
+tempFahrenheit = float(input('Informe a temperatura em Fahrenheit: '))
 tempCelsius = (5/9)*(tempFahrenheit - 32)
 print('Grau Celsius: ',tempCelsius)
